use clickthrough component for VideoPlayer
