+++ common/views/components/VideoPlayer/VideoPlayer.tsx
@@ -10,22 +10,7 @@ import {
 } from '@weco/common/utils/iiif';
 import Space from '@weco/common/views/components/styled/Space';
 import DownloadLink from '@weco/catalogue/components/DownloadLink/DownloadLink';
-import ButtonSolid from '@weco/common/views/components/ButtonSolid/ButtonSolid';
-import { font } from '@weco/common/utils/classnames';
-import styled from 'styled-components';
-import useShowClickthrough from '@weco/common/hooks/useShowClickthrough';
-
-// TODO make own thing and share
-const IframeAuthMessage = styled.iframe`
-  display: none;
-`;
-const iframeId = 'authMessage';
-function reloadAuthIframe(document, id: string) {
-  const authMessageIframe: HTMLIFrameElement = document.getElementById(id);
-  // assigning the iframe src to itself reloads the iframe and refires the window.message event
-  // eslint-disable-next-line no-self-assign
-  authMessageIframe.src = authMessageIframe.src;
-}
+import IIIFClickthrough from '@weco/common/views/components/IIIFClickthrough/IIIFClickthrough';
 
 type Props = {
   video: IIIFMediaElement;
@@ -41,7 +26,6 @@ const VideoPlayer: FunctionComponent<Props> = ({
   const annotation: any = getAnnotationFromMediaElement(video);
   const authService = video && getVideoClickthroughService(video);
   const tokenService = authService && getTokenService(authService);
-  const showClickthrough = useShowClickthrough(authService, tokenService);
 
   function trackViewingTime() {
     trackEvent({
@@ -81,96 +65,53 @@ const VideoPlayer: FunctionComponent<Props> = ({
     },
     isPlaying ? 1000 : null
   );
+
   return (
-    <>
-      {tokenService && origin && (
-        <IframeAuthMessage
-          id={iframeId}
-          src={`${tokenService['@id']}?messageId=1&origin=${origin}`}
-        />
-      )}
-      {showClickthrough ? (
-        <div className={font('hnl', 5)}>
-          {authService?.label && (
-            <h2 className={font('hnm', 4)}>{authService?.label}</h2>
-          )}
-          {authService?.description && (
-            <p
-              dangerouslySetInnerHTML={{
-                __html: authService?.description,
-              }}
-            />
-          )}
-          {authService?.['@id'] && origin && (
-            <Space as="span" h={{ size: 'm', properties: ['margin-right'] }}>
-              <ButtonSolid
-                text="Show the content"
-                clickHandler={() => {
-                  trackEvent({
-                    category: 'ButtonSolidLink',
-                    action: 'follow link "Show the content"',
-                    label: `workId: `, // TODO work id
-                  });
-                  const authServiceWindow = window.open(
-                    `${authService?.['@id'] || ''}?origin=${origin}`
-                  );
-                  authServiceWindow &&
-                    authServiceWindow.addEventListener('unload', function() {
-                      reloadAuthIframe(document, iframeId);
-                    });
-                }}
-              />
-            </Space>
-          )}
-        </div>
-      ) : (
-        <>
-          <video
-            onPlay={() => {
-              setIsPlaying(true);
-              trackEvent({
-                category: 'Video',
-                action: 'play video',
+    <IIIFClickthrough authService={authService} tokenService={tokenService}>
+      <video
+        onPlay={() => {
+          setIsPlaying(true);
+          trackEvent({
+            category: 'Video',
+            action: 'play video',
+            label: video['@id'],
+          });
+        }}
+        onPause={() => {
+          setIsPlaying(false);
+        }}
+        controlsList={!showDownloadOptions ? 'nodownload' : undefined}
+        controls
+        preload="none"
+        poster={video.thumbnail}
+        style={{
+          maxWidth: '100%',
+          maxHeight: '260px',
+          display: 'inline-block',
+        }}
+      >
+        <source src={video['@id']} type={video.format} />
+        {`Sorry, your browser doesn't support embedded video.`}
+      </video>
+      {annotation &&
+        annotation.resource &&
+        annotation.resource.format === 'application/pdf' && (
+          <Space v={{ size: 's', properties: ['margin-top'] }}>
+            <DownloadLink
+              href={annotation.resource['@id']}
+              linkText={`Transcript of ${annotation.resource.label} video`}
+              format={'PDF'}
+              trackingEvent={{
+                category: 'Download link',
+                action: 'follow video annotation link',
                 label: video['@id'],
-              });
-            }}
-            onPause={() => {
-              setIsPlaying(false);
-            }}
-            controlsList={!showDownloadOptions ? 'nodownload' : undefined}
-            controls
-            preload="none"
-            poster={video.thumbnail}
-            style={{
-              maxWidth: '100%',
-              maxHeight: '260px',
-              display: 'inline-block',
-            }}
-          >
-            <source src={video['@id']} type={video.format} />
-            {`Sorry, your browser doesn't support embedded video.`}
-          </video>
-          {annotation &&
-            annotation.resource &&
-            annotation.resource.format === 'application/pdf' && (
-              <Space v={{ size: 's', properties: ['margin-top'] }}>
-                <DownloadLink
-                  href={annotation.resource['@id']}
-                  linkText={`Transcript of ${annotation.resource.label} video`}
-                  format={'PDF'}
-                  trackingEvent={{
-                    category: 'Download link',
-                    action: 'follow video annotation link',
-                    label: video['@id'],
-                  }}
-                  mimeType={annotation.resource.format}
-                  trackingTags={['annotation']}
-                />
-              </Space>
-            )}
-        </>
-      )}
-    </>
+              }}
+              mimeType={annotation.resource.format}
+              trackingTags={['annotation']}
+            />
+          </Space>
+        )}
+    </IIIFClickthrough>
   );
 };
